Added input validation. Checks names if they are empty, checks if numbers are entered, checks in negative or zero numbers are used in BMI calculations
